move functions from utils.py
