+++ project_1444/product/utils.py
@@ -1,9 +1,5 @@
-import uuid
 from decimal import Decimal
-from io import BytesIO
 
-import qrcode
-from django.core.files.base import ContentFile
 from django.db.models import Q
 from django.utils import timezone
 from django.utils.text import slugify
@@ -141,48 +137,3 @@ def get_discounted_price(user, subproduct):
         "old_price": original_price,
         "discount_applied": final_discount,
     }
-
-
-def generate_sku():
-    return f"SKU-{uuid.uuid4().hex[:8].upper()}"
-
-
-def generate_qr_code(product):
-    """Генерує QR-код із `sku` або `ean_13`"""
-    qr_data = product.sku or product.ean_13 or product.name
-    qr = qrcode.make(qr_data)
-    qr_io = BytesIO()
-    qr.save(qr_io, format="PNG")
-    qr_file = ContentFile(qr_io.getvalue(), name=f"qr_{product.sku}.png")
-    return qr_file
-
-
-def generate_subproduct_new_article(instance):
-    ModelClass: object = instance.__class__  # noqa N806
-    last_product = ModelClass.__class__.object.order_by("-id").first()
-    # last_product = SubProducts.objects.order_by("-id").first()
-    if last_product:
-        # Припустимо, що перші два символи - це префікс
-        last_article_number = int(last_product.article[4:])
-        new_article = (
-            f"SBPR{last_article_number + 1:05d}"  # Формат: PR00001, PR00002, ...
-        )
-    else:
-        new_article = "SBPR00001"  # Початковий SKU
-    return new_article
-
-
-def generate_product_new_article(instance):
-    # def generate_sku():
-    ModelClass: object = instance.__class__  # noqa N806
-    last_product = ModelClass.__class__.object.order_by("-id").first()
-    # last_product = Product.objects.order_by("-id").first()
-    if last_product:
-        # Припустимо, що перші два символи - це префікс
-        last_article_number = int(last_product.article[2:])
-        new_article = (
-            f"PR{last_article_number + 1:05d}"  # Формат: PR00001, PR00002, ...
-        )
-    else:
-        new_article = "PR00001"  # Початковий SKU
-    return new_article
